fix: 🐛 move events method from common to specific version
