+++ src/contract_wrappers/modules/wallet/vesting_escrow_wallet_wrapper/__tests__/3.0.0.test.ts
@@ -4,6 +4,7 @@ import {
   VestingEscrowWalletContract_3_0_0,
   ISecurityTokenContract_3_0_0,
   Web3Wrapper,
+  PolyTokenEvents_3_0_0,
 } from '@polymathnetwork/abi-wrappers';
 import { getMockedPolyResponse, MockedCallMethod, MockedSendMethod } from '../../../../../test_utils/mocked_methods';
 import { VestingEscrowWallet_3_0_0 } from '../3.0.0';
@@ -118,6 +119,25 @@ describe('VestingEscrowWalletWrapper', () => {
       // Verifications
       verify(mockedContract.treasuryWallet).once();
       verify(mockedMethod.callAsync()).once();
+    });
+  });
+
+  describe('SubscribeAsync', () => {
+    test('should throw as eventName does not belong to VestingEscrowWallet', async () => {
+      // Mocked parameters
+      const mockedParams = {
+        eventName: PolyTokenEvents_3_0_0.Transfer,
+        indexFilterValues: {},
+        callback: () => {},
+        isVerbose: false,
+      };
+
+      // Real call
+      await expect(target.subscribeAsync(mockedParams)).rejects.toEqual(
+        new Error(
+          `Expected eventName to be one of: 'AddSchedule', 'ModifySchedule', 'RevokeAllSchedules', 'RevokeSchedule', 'DepositTokens', 'SendToTreasury', 'SendTokens', 'AddTemplate', 'RemoveTemplate', 'TreasuryWalletChanged', 'Pause', 'Unpause', encountered: Transfer`,
+        ),
+      );
     });
   });
 });
