+++ src/index.ts
@@ -283,4 +283,19 @@ export {
   Provider,
   Web3Wrapper,
   BigNumber,
+  VestingEscrowWalletContract_3_0_0,
+  VestingEscrowWalletEventArgs_3_0_0,
+  VestingEscrowWalletEvents_3_0_0,
+  VestingEscrowWalletPauseEventArgs_3_0_0,
+  VestingEscrowWalletUnpauseEventArgs_3_0_0,
+  VestingEscrowWalletAddScheduleEventArgs_3_0_0,
+  VestingEscrowWalletModifyScheduleEventArgs_3_0_0,
+  VestingEscrowWalletRevokeAllSchedulesEventArgs_3_0_0,
+  VestingEscrowWalletRevokeScheduleEventArgs_3_0_0,
+  VestingEscrowWalletDepositTokensEventArgs_3_0_0,
+  VestingEscrowWalletSendToTreasuryEventArgs_3_0_0,
+  VestingEscrowWalletSendTokensEventArgs_3_0_0,
+  VestingEscrowWalletAddTemplateEventArgs_3_0_0,
+  VestingEscrowWalletRemoveTemplateEventArgs_3_0_0,
+  VestingEscrowWalletTreasuryWalletChangedEventArgs_3_0_0,
 } from '@polymathnetwork/abi-wrappers';
